requirements.txt
Django==1.11
pytz==2017.2
djangorestframework==3.6.2
django-filter==1.0.2
